+++ src/main/java/com/adiha/EventScheduler/utils/mapper/Constants.java
@@ -1,0 +1,8 @@
+package com.adiha.EventScheduler.utils.mapper;
+
+public class Constants {
+    public static final String CREATION_TIME = "creationTime";
+    public static final String START_TIME = "startTime";
+    public static final String POPULARITY = "popularity";
+
+}
